Further preferences re-write: We don’t need the gigantic clutter of a struct for our preferences, now that we’re using Defaults: trim it down by removing old values. We still need the struct to load the old preference files, though. Moving the struct into a single function. We also don’t have to worry about swapping preferences except for the old preferences: deleting the function SwapPrefs and moving it into the same single function. Get rid of the last of the registration code. Code style changes and improvements.

+++ Files/wds_general/Samples.c
@@ -4749,7 +4749,7 @@ Point		theCell = { 0, 0}, ppt = { 0,0};
 				if (thePrefs.PianoKey[ (short) theChar] == MADDriver->pitchTable[ i * MAXSAMPLE +  0] ) break;
 			}*/
 			
-			i = thePrefs.PianoKey[ (short) theChar];
+			i = thePrefs.PianoKey[theChar];
 			
 			if (i != 0xFF && i != -1) i += thePrefs.pianoOffset*12;
